refactor: moved getBatchId to common utils
