+++ src/account/utils.ts
@@ -50,19 +50,3 @@ export function assertActiveAccount(data: AccountData): void {
     throw new Error('Active account not found')
   }
 }
-
-export async function getBatchId(beeDebug: BeeDebug): Promise<string> {
-  const batches = await beeDebug.getAllPostageBatch()
-
-  if (batches.length === 0) {
-    throw new Error('Postage batch not exists')
-  }
-
-  const batchId = batches.pop()?.batchID
-
-  if (!batchId) {
-    throw new Error('Incorrect batch id found')
-  }
-
-  return batchId
-}
